ajeitando eventos

+++ database/migrations/2020_11_20_120348_create_medida_table.php
@@ -23,10 +23,10 @@ class CreateMedidaTable extends Migration
             $table->string('tipo_noticia')->nullable();
             $table->text('texto_noticia')->nullable();
             $table->text('titulo_noticia')->nullable();
-            $table->bigInteger('diferenca_financas')->nullable();
-            $table->bigInteger('diferenca_popularidade_empresarios')->nullable();
-            $table->bigInteger('diferenca_popularidade_trabalhadores')->nullable();
-            $table->bigInteger('diferenca_popularidade_estado')->nullable();
+            $table->decimal('diferenca_financas', 13, 3)->nullable();
+            $table->decimal('diferenca_popularidade_empresarios', 4, 3)->nullable();
+            $table->decimal('diferenca_popularidade_trabalhadores', 4, 3)->nullable();
+            $table->decimal('diferenca_popularidade_estado', 4, 3)->nullable();
             $table->timestamps();
         });
     }
